feat(core): Test coupon application at checkout
